+++ processor/src/chiplets/ace/mod.rs
@@ -10,21 +10,20 @@ use vm_core::{FieldElement, ZERO, mast::BasicBlockNode};
 
 use crate::{
     ContextId, ExecutionError, Felt, QuadFelt,
-    chiplets::{ace::trace::CircuitEvaluation, memory::Memory},
+    chiplets::memory::Memory,
     errors::{AceError, ErrorContext},
-    fast::MemoryFast,
     trace::TraceFragment,
 };
 
 mod trace;
-pub use trace::{NUM_ACE_LOGUP_FRACTIONS_EVAL, NUM_ACE_LOGUP_FRACTIONS_READ};
+pub use trace::{CircuitEvaluation, NUM_ACE_LOGUP_FRACTIONS_EVAL, NUM_ACE_LOGUP_FRACTIONS_READ};
 
 mod instruction;
 #[cfg(test)]
 mod tests;
 
-const PTR_OFFSET_ELEM: Felt = Felt::ONE;
-const PTR_OFFSET_WORD: Felt = Felt::new(4);
+pub const PTR_OFFSET_ELEM: Felt = Felt::ONE;
+pub const PTR_OFFSET_WORD: Felt = Felt::new(4);
 
 /// Arithmetic circuit evaluation (ACE) chiplet.
 ///
@@ -77,8 +76,12 @@ impl Ace {
     }
 
     /// Adds an entry resulting from a call to the ACE chiplet.
-    pub(crate) fn add_eval_context(&mut self, clk: RowIndex, eval_context: CircuitEvaluation) {
-        self.circuit_evaluations.insert(clk, eval_context);
+    pub(crate) fn add_circuit_evaluation(
+        &mut self,
+        clk: RowIndex,
+        circuit_eval: CircuitEvaluation,
+    ) {
+        self.circuit_evaluations.insert(clk, circuit_eval);
     }
 }
 
@@ -330,69 +333,6 @@ pub fn eval_circuit(
     for _ in 0..num_eval_rows {
         let instruction =
             mem.read(ctx, ptr, clk, error_ctx).map_err(ExecutionError::MemoryError)?;
-        evaluation_context.do_eval(ptr, instruction, error_ctx)?;
-        ptr += PTR_OFFSET_ELEM;
-    }
-
-    // Ensure the circuit evaluated to zero.
-    if !evaluation_context.output_value().is_some_and(|eval| eval == QuadFelt::ZERO) {
-        return Err(ExecutionError::failed_arithmetic_evaluation(
-            error_ctx,
-            AceError::CircuitNotEvaluateZero,
-        ));
-    }
-
-    Ok(evaluation_context)
-}
-
-/// Identical to `[eval_circuit]` but addapted for use with `[FastProcessor]`.
-#[allow(clippy::too_many_arguments)]
-pub fn eval_circuit_fast(
-    ctx: ContextId,
-    ptr: Felt,
-    clk: RowIndex,
-    num_vars: Felt,
-    num_eval: Felt,
-    mem: &mut MemoryFast,
-    op_idx: usize,
-    error_ctx: &ErrorContext<'_, BasicBlockNode>,
-) -> Result<CircuitEvaluation, ExecutionError> {
-    let num_vars = num_vars.as_int();
-    let num_eval = num_eval.as_int();
-
-    // Ensure vars and instructions are word-aligned and non-empty. Note that variables are
-    // quadratic extension field elements while instructions are encoded as base field elements.
-    // Hence we can pack 2 variables and 4 instructions per word.
-    if num_vars % 2 != 0 || num_vars == 0 {
-        return Err(ExecutionError::failed_arithmetic_evaluation(
-            error_ctx,
-            AceError::NumVarIsNotWordAlignedOrIsEmpty(num_vars),
-        ));
-    }
-    if num_eval % 4 != 0 || num_eval == 0 {
-        return Err(ExecutionError::failed_arithmetic_evaluation(
-            error_ctx,
-            AceError::NumEvalIsNotWordAlignedOrIsEmpty(num_eval),
-        ));
-    }
-
-    // Ensure instructions are word-aligned and non-empty
-    let num_read_rows = num_vars as u32 / 2;
-    let num_eval_rows = num_eval as u32;
-
-    let mut evaluation_context =
-        CircuitEvaluation::new(ctx, clk + op_idx, num_read_rows, num_eval_rows);
-
-    let mut ptr = ptr;
-    // perform READ operations
-    for _ in 0..num_read_rows {
-        let word = mem.read_word(ctx, ptr, clk + op_idx)?;
-        evaluation_context.do_read(ptr, *word)?;
-        ptr += PTR_OFFSET_WORD;
-    }
-    // perform EVAL operations
-    for _ in 0..num_eval_rows {
-        let instruction = mem.read_element(ctx, ptr)?;
         evaluation_context.do_eval(ptr, instruction, error_ctx)?;
         ptr += PTR_OFFSET_ELEM;
     }
